fix(migrations): Fix merkle_batch_id type in activity_log_archive migration
- Changed integer to bigInteger for merkle_batch_id column
- Matches original migration schema and test expectations
- ActivityLogSchemaTest now passes (verifies bigint type)
- Added RefreshDatabase trait to ActivityLogSchemaTest for CI compatibility

Fixes failing CI tests in PR #436

+++ database/migrations/2025_12_27_040629_change_activity_log_archive_id_to_bigint.php
@@ -50,7 +50,7 @@ return new class extends Migration
 
             // 🌳 Merkle Tree Metadata
             $table->string('merkle_root', 64)->nullable();
-            $table->integer('merkle_batch_id')->nullable();
+            $table->bigInteger('merkle_batch_id')->nullable();
         });
     }
 
@@ -72,7 +72,7 @@ return new class extends Migration
             $table->string('event_hash', 64)->nullable();
             $table->string('previous_hash', 64)->nullable();
             $table->string('merkle_root', 64)->nullable();
-            $table->integer('merkle_batch_id')->nullable();
+            $table->bigInteger('merkle_batch_id')->nullable();
         });
     }
 };
